KeyCeremony hashing conform to 2.0

+++ egklib/src/commonMain/kotlin/electionguard/keyceremony/KeyCeremonyTrustee.kt
@@ -199,47 +199,51 @@ open class KeyCeremonyTrustee(
 
         val K_l = other.publicKey() // other's publicKey
         val hp = K_l.context.constants.hp.bytes
-        val i = xCoordinate.toUShort()
-        val l = other.guardianXCoordinate.toUShort()
+        val i = xCoordinate
+        val l = other.guardianXCoordinate
 
-        // (alpha, beta) = (g^R mod p, K^R mod p)  spec 2.0.0, eq 14
+        // (αi,ℓ , βi,ℓ ) = (g^ξi,ℓ mod p, K^ξi,ℓ mod p), ξi,ℓ = nonce
+        // (α_i,ℓ , β_i,ℓ ) = (g^nonce mod p, Kℓ^nonce mod p) ;  spec 2.0.0, eq 14
         // by encrypting a zero, we achieve exactly this
         val (alpha, beta) = 0.encrypt(K_l, nonce)
-        // ki,ℓ = H(HP ; 11, i, ℓ, Kℓ , αi,ℓ , βi,ℓ ) spec 2.0.0, eq 15
+        // ki,ℓ = H(HP ; 0x11, i, ℓ, Kℓ , αi,ℓ , βi,ℓ ) ; eq 15 "secret key"
         val kil = hashFunction(hp, 0x11.toByte(), i, l, K_l.key, alpha, beta).bytes
 
         // footnote 27
         // This key derivation uses the KDF in counter mode from SP 800-108r1. The second input to HMAC contains
-        // the counter in the first byte, the UTF-8 encoding of the string "share enc keys" as the Label (encoding is denoted
-        // by b(. . . ), see Section 5.1.4), a separation 00 byte, the UTF-8 encoding of the string "share encrypt" concatenated
+        // the counter in the first byte, the UTF-8 encoding of the string "share_enc_keys" as the Label (encoding is denoted
+        // by b(. . . ), see Section 5.1.4), a separation 00 byte, the UTF-8 encoding of the string "share_encrypt" concatenated
         // with encodings of the numbers i and ℓ of the sending and receiving guardians as the Context, and the final two bytes
         // specifying the length of the output key material as 512 bits in total.
 
-        // context = b(”share encrypt”) ∥ b(i, 2) ∥ b(ℓ, 2)
-        // k0 = HMAC(ki,ℓ , 01 ∥ label ∥ 00 ∥ context ∥ 0200) spec 2.0.0,  eq 16
-        val k0 = hmacFunction(kil, 0x01.toByte(), label, 0x00.toByte(), context, i, l, 512.toShort()).bytes
-        // k1 = HMAC(ki,ℓ , 02 ∥ label ∥ 00 ∥ context ∥ 0200), spec 2.0.0, eq 17
-        val k1 = hmacFunction(kil, 0x02.toByte(), label, 0x00.toByte(), context, i, l, 512.toShort()).bytes
+        // Label = b(“share enc keys”, 14)
+        // Context = b(“share_encrypt”, 13) ∥ b(i, 4) ∥ b(ℓ, 4)
+        // k0 = HMAC(ki,ℓ , 0x01 ∥ Label ∥ 0x00 ∥ Context ∥ 0x0200) ; spec 2.0.0,  eq 16
+        val k0 = hmacFunction(kil, 0x01.toByte(), label, 0x00.toByte(), context, i, l, 512).bytes
+        // k1 = HMAC(ki,ℓ , 0x02 ∥ Label ∥ 0x00 ∥ Context ∥ 0x0200) ; eq 17
+        val k1 = hmacFunction(kil, 0x02.toByte(), label, 0x00.toByte(), context, i, l, 512).bytes
 
         // spec 2.0.0, eq 19
-        // C0 = g^nonce == alpha
+        // Ci,ℓ,0 = g^ξi,ℓ mod p = C0 = g^nonce == alpha
         val c0: ElementModP = alpha
-        // C1 = b(Pi(ℓ),32) ⊕ k1, • The symbol ⊕ denotes bitwise XOR.
+        // C1 = b(Pi(ℓ),32) ⊕ k1, where the symbol ⊕ denotes bitwise XOR.
         val pilBytes = Pil.byteArray()
         val c1 = ByteArray(32) { pilBytes[it] xor k1[it] }
-        // C2 = HMAC(k0 , b(Ci,ℓ,0 , 512) ∥ Ci,ℓ,1 )
+        // Ci,ℓ,2 = HMAC(k0 , b(Ci,ℓ,0 , 512) ∥ Ci,ℓ,1 )
+        // C2 = HMAC(k0, b(C0,512) ∥ C1 )
         val c2 = hmacFunction(k0, c0, c1)
 
         return HashedElGamalCiphertext(c0, c1, c2, pilBytes.size)
     }
 
-    // Share decryption. After receiving the ciphertext (Ci,ℓ,0 , Ci,ℓ,1 , Ci,ℓ,2 ) from guardian Gi , guardian
-    // Gℓ decrypts it by computing βi,ℓ = (Ci,ℓ,0 )sℓ mod p, setting αi,ℓ = Ci,ℓ,0 and obtaining
+    // Share decryption.
+    // After receiving the ciphertext (Ci,ℓ,0 , Ci,ℓ,1 , Ci,ℓ,2 ) from guardian Gi , guardian
+    // Gℓ decrypts it by computing βi,ℓ = (Ci,ℓ,0)^sℓ mod p, setting αi,ℓ = Ci,ℓ,0 and obtaining
     //   ki,ℓ = H(HP ; 11, i, ℓ, Kℓ , αi,ℓ , βi,ℓ ).
     // Now the MAC key k0 and the encryption key k1 can be computed as
     // above in Equations (16) and (17), which allows Gℓ to verify the validity of the MAC, namely that
     //   Ci,ℓ,2 = HMAC(k0 , b(Ci,ℓ,0 , 512) ∥ Ci,ℓ,1 ).
-    // If the MAC verifies, then Gℓ decrypts b(Pi (ℓ), 32) = Ci,ℓ,1 ⊕k1 .
+    // If the MAC verifies, then Gℓ decrypts b(Pi (ℓ), 32) = Ci,ℓ,1 ⊕ k1 .
     fun shareDecryption(share: EncryptedKeyShare): ByteArray?  {
         // αi,ℓ = Ci,ℓ,0
         // βi,ℓ = (Ci,ℓ,0 )sℓ mod p
@@ -250,23 +254,19 @@ open class KeyCeremonyTrustee(
         val alpha = c0
         val beta = c0 powP this.electionPrivateKey()
         val hp = group.constants.hp.bytes
-        val kil = hashFunction(hp, 0x11.toByte(), share.ownerXcoord.toShort(), xCoordinate.toShort(), electionPublicKey(), alpha, beta).bytes
+        val kil = hashFunction(hp, 0x11.toByte(), share.ownerXcoord, xCoordinate, electionPublicKey(), alpha, beta).bytes
 
         // Now the MAC key k0 and the encryption key k1 can be computed as above in Equations (16) and (17)
-        // context = b(”share encrypt”) ∥ b(i, 2) ∥ b(ℓ, 2)
-        // k0 = HMAC(ki,ℓ , 01 ∥ label ∥ 00 ∥ context ∥ 0200) eq 15
-        val k0 = hmacFunction(kil, 0x01.toByte(), label, 0x00.toByte(), context, share.ownerXcoord.toShort(), xCoordinate.toShort(), 512.toShort()).bytes
-        // k1 = HMAC(ki,ℓ , 02 ∥ label ∥ 00 ∥ context ∥ 0200), eq 16
-        val k1 = hmacFunction(kil, 0x02.toByte(), label, 0x00.toByte(), context, share.ownerXcoord.toShort(), xCoordinate.toShort(), 512.toShort()).bytes
+        val k0 = hmacFunction(kil, 0x01.toByte(), label, 0x00.toByte(), context, share.ownerXcoord, xCoordinate, 512).bytes
+        val k1 = hmacFunction(kil, 0x02.toByte(), label, 0x00.toByte(), context, share.ownerXcoord, xCoordinate, 512).bytes
 
-        // Gℓ can verify the validity of the MAC, namely that
-        // Ci,ℓ,2 = HMAC(k0 , b(Ci,ℓ,0 , 512) ∥ Ci,ℓ,1 ). If the MAC verifies, Gℓ decrypts b(Pi (ℓ), 32) = Ci,ℓ,1 ⊕ k1 .
+        // Gℓ can verify the validity of the MAC, namely that Ci,ℓ,2 = HMAC(k0 , b(Ci,ℓ,0 , 512) ∥ Ci,ℓ,1 ).
         val expectedC2 = hmacFunction(k0, c0, c1)
         if (expectedC2 != share.encryptedCoordinate.c2) {
             return null
         }
 
-        //  If the MAC verifies, Gℓ decrypts b(Pi (ℓ), 32) = Ci,ℓ,1 ⊕ k1 .
+        //  If the MAC verifies, Gℓ decrypts b(Pi(ℓ), 32) = Ci,ℓ,1 ⊕ k1 .
         val pilBytes = ByteArray(32) { c1[it] xor k1[it] }
         return pilBytes
     }
